cpbc-volunteer-app/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
boto3>=1.33.0
python-dotenv>=1.0.0
alembic>=1.12.0
pydantic[email]>=2.5.0
